Add "real" logging library
The current logger is a light wrapper around `print` and `inspect`.
To top it off this doesn't seem to work on windows.
This is the start of a refactor that will remove the current logger and
replace it with a MIT licensed logger here: https://github.com/rxi/log.lua